Moves auxiliary function to separate module.
withinRange is a function that returns the closest number within
a certain range. I keep writing this function again and again across
programs.

This commit introduces a Data.Extra.Ord module with that function,
and modifies the code that uses it.

The cabal file is changed accordingly.

Also: The other-modules field is sorted alphabetically.