Re-sort fields (KSG Cabal V1.1.0 - 3.7.1). Refs #11.
The fields inside executables and test benches must follow a spcific order.
With respect to these changes, the order must be:

- exposed-modules
- build-depends
- hs-source-dirs
- compiler options

In this file, many options are under an os(windows) check. I am choosing to
treat those as flags, and put them after normal compiler options. It might be
possible that they can be written unconditionally, with only small additions
for Windows.

Also, I am treating extra-libraries as a compiler-level option and not a normal
dependency. Our Cabal style guide does not specify where it should go.